# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
  /api/users/me:
    get:
      tags:
      - user-rest-controller
      summary: Shows the logged user's profile
      operationId: getUser
      parameters:
      - name: pageItineraries
        in: query
        description: itineraries page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: pageReviews
        in: query
        description: reviews page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "400":
          description: Operation not available when no user is logged in
        "200":
          description: Displayed profile properly
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
      - user-rest-controller
      summary: Edits the user's information
      operationId: editUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "400":
          description: Email does not match against correct mail format
        "403":
          description: No user is logged in
        "200":
          description: Successfully edited user's data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "500":
          description: Empty body of request
...
  /api/management/users/{id}:
    get:
      tags:
      - user-management-rest-controller
      summary: Get the specified user
      operationId: getUser_1
      parameters:
      - name: id
        in: path
        description: id of user
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: Not authorized
        "404":
          description: User not found
    put:
      tags:
      - user-management-rest-controller
      summary: Edit existing user
      operationId: editUser_1
      parameters:
      - name: id
        in: path
        description: id of user to be edited
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        "200":
          description: User was correctly edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "403":
          description: Not authorized
        "404":
          description: User not found
    delete:
      tags:
      - user-management-rest-controller
      summary: Delete user
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: id of user to be deleted
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: User was correctly deleted
        "403":
          description: Not authorized
        "404":
          description: User not found
...
  /api/management/itineraries/{id}:
    put:
      tags:
      - itinerary-management-rest-controller
      summary: Edit itinerary
      operationId: editItinerary
      parameters:
      - name: id
        in: path
        description: edited item
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryDTO'
        required: true
      responses:
        "404":
          description: Destination not found
        "200":
          description: Itinerary is edited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerary'
        "403":
          description: Not authorized
    delete:
      tags:
      - itinerary-management-rest-controller
      summary: Delete itinerary by id
      operationId: deleteItinerary
      parameters:
      - name: id
        in: path
        description: delete itinerary by id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: itinerary is deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerary'
        "404":
          description: Destination not found
        "403":
          description: Not authorized
...
  /api/itineraries/{id}:
    get:
      tags:
      - itinerary-rest-controller
      summary: Get Detailed information about a specific Itinerary.
      operationId: itinerary
      parameters:
      - name: id
        in: path
        description: itinerary id
        required: true
        schema:
          type: integer
          format: int32
      - name: placesPage
        in: query
        description: places page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: reviewsPage
        in: query
        description: reviews page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "404":
          description: Requested a non-existing Itinerary.
        "403":
          description: Invalid permissions to request places from a private Itinerary.
        "200":
          description: Successfully searched the desired Itineraries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItineraryDetailsDTO'
        "400":
          description: Invalid arguments.
    put:
      tags:
      - itinerary-rest-controller
      summary: Edit an itinerary given its id
      operationId: editItineraryById
      parameters:
      - name: id
        in: path
        description: id of the itinerary to edit
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryDTO'
        required: true
      responses:
        "403":
          description: Login needed to perform operation or User is not owner of the
            itinerary
        "200":
          description: Itinerary edited successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerary'
        "404":
          description: Itinerary not found
    delete:
      tags:
      - itinerary-rest-controller
      summary: Delete an itinerary given its id
      operationId: deleteItinerary_1
      parameters:
      - name: id
        in: path
        description: id of the itinerary to be deleted
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successfully deleted the itinerary
        "403":
          description: Operation not supported for not registered users
        "404":
          description: Itinerary not found
  /api/itineraries/{id}/image:
    get:
      tags:
      - itinerary-rest-controller
      summary: Returns the image of the desired Itinerary.
      operationId: downloadImage_2
      parameters:
      - name: id
        in: path
        description: itinerary id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Requested a non-existing Itinerary's image.
        "200":
          description: Sucessfully returned the Itinerary's image.
          content:
            image/jpeg: {}
    put:
      tags:
      - itinerary-rest-controller
      summary: Edit the image of an itinerary
      operationId: editItineraryImage
      parameters:
      - name: id
        in: path
        description: id of the itinerary you want to add an image to
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
          image/jpeg:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
          image/png:
            schema:
              required:
              - imageFile
              type: object
              properties:
                imageFile:
                  type: string
                  description: image to set as itinerary image
                  format: binary
      responses:
        "403":
          description: Login needed to perform operation or User is not owner of the
            itinerary
        "404":
          description: Itinerary not found
        "200":
          description: Image edited successfully
          content:
            image/jpeg: {}
...
  /api/management/places:
    get:
      tags:
      - place-management-rest-controller
      summary: Get a page with all places
      operationId: getPlaces
      parameters:
      - name: page
        in: query
        description: page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "404":
          description: Places not found
        "200":
          description: Places page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "403":
          description: Not authorized
    post:
      tags:
      - place-management-rest-controller
      summary: Create a new place
      operationId: createPlace
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceDTO'
        required: true
      responses:
        "201":
          description: New place was created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Place'
        "403":
          description: Not authorized
...
  /api/itineraries:
    get:
      tags:
      - itinerary-rest-controller
      summary: Searches for Itineraries in the database. Filters can be applied optionally.
      operationId: getItineraries
      parameters:
      - name: name
        in: query
        description: search query
        required: false
        schema:
          type: string
          default: ""
      - name: sort
        in: query
        description: "sorting type: id, name, views"
        required: false
        schema:
          type: string
          default: id
      - name: order
        in: query
        description: order specifier
        required: false
        schema:
          type: string
          default: DESC
      - name: page
        in: query
        description: page number
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "200":
          description: Successfully searched the desired Itinerary.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerary'
        "400":
          description: Invalid arguments.
    post:
      tags:
      - itinerary-rest-controller
      summary: "Creates a new itinerary if itiDTO is present. If it is not present\
        \ and there is a copyFrom id, copies the itinerary with id = copyFrom"
      operationId: createOrCopyItinerary
      parameters:
      - name: copyFrom
        in: query
        description: id of the itinerary to copy
        required: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItineraryDTO'
      responses:
        "400":
          description: itiDTO data missing for POSTing new itinerary
        "201":
          description: Newly created itinerary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Itinerary'
        "403":
          description: Operation not supported for users without an account
        "404":
          description: Itinerary to copy from not found
...
  /api/users/{id}/image:
    get:
      tags:
      - user-rest-controller
      summary: Returns the profile image of the desired user
      operationId: downloadImage
      parameters:
      - name: id
        in: path
        description: user id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successfully returned the user image
          content:
            image/jpeg: {}
        "404":
          description: Requested a non-existing User's image.
...
  /api/places/{id}/image:
    get:
      tags:
      - place-rest-controller
      summary: Returns the image of the desired Place.
      operationId: downloadImage_1
      parameters:
      - name: id
        in: path
        description: place id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Requested a non-existing Place's image.
        "200":
          description: Sucessfully returned the Place's image.
          content:
            image/jpeg: {}
...
  /api/destinations/{id}/image:
    get:
      tags:
      - destination-rest-controller
      summary: Returns the image of the desired Destination.
      operationId: downloadImage_3
      parameters:
      - name: id
        in: path
        description: destination id
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successfully returned the Destination's image.
          content:
            image/jpeg: {}
        "404":
          description: Requested a non-existing Destination's image.
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        nationality:
          type: string
        image:
          type: boolean
        roles:
          type: array
          items:
            type: string
    UserDTO:
      type: object
      properties:
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        passwordHash:
          type: string
        nationality:
          type: string
      description: new user's information
    PlaceDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        destination:
          type: string
      description: new page's information
    Destination:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        views:
          type: integer
          format: int64
        image:
          type: boolean
        flagCode:
          type: string
        type:
          type: string
        flag:
          type: string
        typeLowercase:
          type: string
    Place:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        views:
          type: integer
          format: int64
        image:
          type: boolean
        destination:
          $ref: '#/components/schemas/Destination'
        type:
          type: string
        flag:
          type: string
        typeLowercase:
          type: string
    ItineraryDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        publicValue:
          type: boolean
          writeOnly: true
        user:
          type: string
        public:
          type: boolean
      description: data of the itinerary to create
    Itinerary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        views:
          type: integer
          format: int64
        image:
          type: boolean
        user:
          $ref: '#/components/schemas/User'
        public:
          type: boolean
        type:
          type: string
        flag:
          type: string
        typeLowercase:
          type: string
...
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
...
    Sort:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    ItineraryDetailsDTO:
      type: object
      properties:
        itinerary:
          $ref: '#/components/schemas/Itinerary'
        places:
          $ref: '#/components/schemas/PagePlace'
        reviews:
          $ref: '#/components/schemas/PageReview'
    PagePlace:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Place'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    PageReview:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        views:
          type: integer
          format: int64
        score:
          type: integer
          format: int32
        date:
          type: string
          format: date-time
        itinerary:
          $ref: '#/components/schemas/Itinerary'
        user:
          $ref: '#/components/schemas/User'
        dateToString:
          type: string
...